+++ Source/Gapotchenko.FX.Console/Emulation/IVirtualTerminalOutputBackend.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.IO;
+
+namespace Gapotchenko.FX.Console.Emulation
+{
+    interface IVirtualTerminalOutputBackend
+    {
+        TextWriter Out { get; }
+        ConsoleColor ForegroundColor { get; set; }
+        ConsoleColor BackgroundColor { get; set; }
+
+        void ResetColor();
+        void ResetForegroundColor();
+        void ResetBackgroundColor();
+    }
+}