+++ Documentation/Changelog.md
@@ -7,7 +7,8 @@
 Release date: not released yet
 
 - Introduced `Gapotchenko.FX.AppModel.Information` module that allows to programmatically retrieve information about the app
-- Added `Process.GetImageFileName()` method that allows to retrieve file name of a running process without security restrictions imposed by OS
+- Introduced `Gapotchenko.FX.Console` module that provides virtual terminal operations, console traits and a powerful `MoreTextWriter` primitive for improved user friendliness of your console apps
+- Added `Process.GetImageFileName()` method that allows to retrieve file name of a running process without security restrictions imposed by the host OS
 - .NET Framework 4.0 target is retired. The minimal supported version of .NET Framework is 4.5
 - Fixed issue with ambiguous match of `IsNullOrEmpty` polyfill method of `HashSet<T>` type that occurred in .NET 4.6+, .NET Standard 2.0+ and .NET Core 2.0+ target frameworks
 - Fixed issue with ambiguous match of `ToHashSet` polyfill method of `IEnumerable<T>` type that occurred in .NET 4.7.2+ target frameworks